requirements.txt
absl-py==1.2.0
anyio==3.6.1
appdirs==1.4.4
argcomplete==1.10.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
async-generator==1.10
attrs==22.1.0
autopep8==1.7.0
Babel==2.10.3
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.11.1
biopython>=1.79
black==22.8.0
bleach==5.0.1
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
compressed-rtf==1.0.6
contourpy==1.0.5
cryptography==35.0.0
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
docx2txt==0.8
ebcdic==1.1.1
entrypoints==0.4
exceptiongroup==1.0.0rc9
executing==1.1.0
extract-msg==0.28.7
fastjsonschema==2.16.2
flit_core==3.7.1
fonttools==4.37.4
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.10.2
google-api-python-client==2.64.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.56.4
grpcio==1.49.1
h11==0.14.0
idna==3.4
IMAPClient==2.1.0
importlib-metadata==4.11.4
importlib-resources==5.9.0
ipykernel==6.16.0
ipython~=8.5.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.2
joblib~=1.2.0
json5==0.9.5
jsonschema==4.16.0
jupyter-client==7.3.4
jupyter-contrib-core==0.4.0
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.11.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.5.0
jupyter-server==1.18.1
jupyter-server-mathjax==0.2.6
jupyterlab==3.4.4
jupyterlab-autorun-cells==1.0.0
jupyterlab-cell-flash==0.3.5
jupyterlab-code-formatter==1.5.3
jupyterlab-git==0.39.3
jupyterlab-pygments==0.1.2
jupyterlab_server==2.15.2
jupyterlab-templates==0.3.1
jupyterlab-widgets==1.0.0
kiwisolver==1.4.4
lckr-jupyterlab-variableinspector==3.0.9
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib~=3.6.0
matplotlib-inline==0.1.6
mistune==2.0.3
mypy-extensions==0.4.3
nbclassic==0.4.4
nbclient==0.5.13
nbconvert==7.1.0
nbdime==3.1.1
nbformat==5.6.1
nest-asyncio==1.5.6
nltk==3.7
notebook==6.4.12
notebook-shim==0.1.0
numpy~=1.23.3
oauth2client==4.1.3
oauthlib==3.2.1
olefile==0.46
outcome==1.2.0
packaging==21.3
pandas~=1.5.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.1
pdfminer.six==20191110
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.2.2
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
plotly==5.10.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pycryptodome==3.15.0
PyDrive==1.3.1
pyee==8.2.2
Pygments==2.13.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-pptx==0.6.21
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==23.2.0
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
seaborn~=0.12.0
selenium~=4.5.0
Send2Trash==1.8.0
setuptools~=63.4.1
six==1.16.0
smmap==3.0.5
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
stack-data==0.5.1
tenacity==8.0.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.16.0
testpath==0.6.0
theme-darcula==4.0.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tornado==6.1
tqdm==4.64.1
traitlets==5.4.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.11
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
websockets==10.3
Werkzeug==2.2.2
wheel==0.37.1
widgetsnbextension==3.5.2
wsproto==1.2.0
xlrd==1.2.0
XlsxWriter==3.0.3
zipp==3.8.1

textract
pymysql~=1.0.2
sqlalchemy~=1.4.42
colorcet~=3.0.1
adjusttext~=0.7.3
cairosvg~=2.5.2